Homolog

+++ modernize/src/app/(DashboardLayout)/apps/producao/arte-final/page.tsx
@@ -80,8 +80,10 @@ const ArteFinalScreen = () => {
   const accessToken = localStorage.getItem('accessToken');
   const designers = localStorage.getItem('designers');
   const roles = localStorage.getItem('roles')?.split(',').map(Number) || [];
-  const allowedRoles = [1, 2, 3, 4, 7]; // colocar backoffice
-  const DesignerRoles = [6, 7];
+  const allowedRoles = [1, 2, 4, 7, 10]; // colocar backoffice
+  const DesignerRoles = [6];
+  const DesignerCoordanateRole = [7];
+  const canShowButtonAtribuirDesigner = roles.some(role => DesignerCoordanateRole.includes(role));
   const canShowButton = roles.some(role => allowedRoles.includes(role));
   const canShowButtonDesigner = roles.some(role => DesignerRoles.includes(role));
 
@@ -129,12 +131,19 @@ const ArteFinalScreen = () => {
   };
 
   const handleDelete = async (row: ArteFinal) => {
-    const sucesso = await deletePedidoArteFinal(row?.id, refetch);
-    if (sucesso) {
-      console.log("Pedido deletado com sucesso!");
-      alert('sucesso');
+    const confirmar = window.confirm('Deseja excluir o pedido N° ' + row.numero_pedido);
+    if (confirmar) {
+      const sucesso = await deletePedidoArteFinal(row?.id, refetch);
+      if (sucesso) {
+        console.log("Pedido deletado com sucesso!");
+        alert('sucesso');
+      } else {
+        alert('Falha ao excluir pedido.');
+        console.log("Falha ao excluir pedido.");
+      }
     } else {
-      console.log("Falha ao excluir pedido.");
+      console.log("Envio cancelado.");
+      alert('Envio cancelado.');
     }
   };
 
@@ -528,7 +537,6 @@ const ArteFinalScreen = () => {
                           ) : (
                             <TableCell align="center">
                               {canShowButtonDesigner ? (
-
                                 <Button
                                   sx={{
                                     backgroundColor: 'transparent',
@@ -667,15 +675,19 @@ const ArteFinalScreen = () => {
                                 <IconPrinter />
                               </IconButton>
                             </Tooltip>
-                            {/* validar a role do user 3,4,1*/}
 
+                            {/* validar a role do user 6*/}
+                            {canShowButtonAtribuirDesigner && (
+                              <Tooltip title="Atribuir Designer">
+                                <IconButton onClick={() => handleAtribuirDesigner(row)}>
+                                  <IconBrush />
+                                </IconButton>
+                              </Tooltip>
+                            )}
+
+                            {/* validar a role do user 2,4,1*/}
                             {canShowButton && (
                               <>
-                                <Tooltip title="Atribuir Designer">
-                                  <IconButton onClick={() => handleAtribuirDesigner(row)}>
-                                    <IconBrush />
-                                  </IconButton>
-                                </Tooltip>
                                 <Tooltip title="Editar">
                                   <IconButton onClick={() => handleEdit(row)}>
                                     <IconEdit />
